[wgpu-hal]: MVP `no_std` support (#7599)
* MVP `no_std` support in `wgpu-hal`

* Update CHANGELOG.md

* Fix visibility

* Fix unused imports

* Response to feedback


* Update other `validation_canary` usages


---------
