+++ CHANGELOG.md
@@ -106,6 +106,10 @@ Naga now infers the correct binding layout when a resource appears only in an as
 
 - The type of the `size` parameter to `copy_buffer_to_buffer` has changed from `BufferAddress` to `impl Into<Option<BufferAddress>>`. This achieves the spec-defined behavior of the value being optional, while still accepting existing calls without changes. By @andyleiserson in [#7659](https://github.com/gfx-rs/wgpu/pull/7659).
 
+#### HAL
+
+- Added initial `no_std` support to `wgpu-hal`. By @bushrat011899 in [#7599](https://github.com/gfx-rs/wgpu/pull/7599)
+
 ### Bug Fixes
 
 #### Naga
