+++ tests/Cargo.toml
@@ -32,6 +32,7 @@ webgl = ["wgpu/webgl"]
 
 [dependencies]
 wgpu = { workspace = true, features = ["noop"] }
+wgpu-hal = { workspace = true, features = ["validation_canary"] }
 wgpu-macros.workspace = true
 
 anyhow.workspace = true
